Move _clear_error_support.py -> _utils/clear_error_support.py
