+++ cuda_core/cuda/core/experimental/_device.py
@@ -5,10 +5,10 @@
 import threading
 from typing import Union
 
-from cuda.core.experimental._clear_error_support import assert_type
 from cuda.core.experimental._context import Context, ContextOptions
 from cuda.core.experimental._memory import Buffer, MemoryResource, _DefaultAsyncMempool, _SynchronousMemoryResource
 from cuda.core.experimental._stream import Stream, StreamOptions, default_stream
+from cuda.core.experimental._utils.clear_error_support import assert_type
 from cuda.core.experimental._utils.cuda_utils import (
     ComputeCapability,
     CUDAError,
