+++ cuda_core/cuda/core/experimental/_linker.py
@@ -14,9 +14,9 @@ from warnings import warn
 if TYPE_CHECKING:
     import cuda.bindings
 
-from cuda.core.experimental._clear_error_support import assert_type
 from cuda.core.experimental._device import Device
 from cuda.core.experimental._module import ObjectCode
+from cuda.core.experimental._utils.clear_error_support import assert_type
 from cuda.core.experimental._utils.cuda_utils import check_or_create_options, driver, handle_return, is_sequence
 
 # TODO: revisit this treatment for py313t builds
