+++ cuda_core/cuda/core/experimental/_program.py
@@ -12,10 +12,10 @@ from warnings import warn
 if TYPE_CHECKING:
     import cuda.bindings
 
-from cuda.core.experimental._clear_error_support import assert_type
 from cuda.core.experimental._device import Device
 from cuda.core.experimental._linker import Linker, LinkerHandleT, LinkerOptions
 from cuda.core.experimental._module import ObjectCode
+from cuda.core.experimental._utils.clear_error_support import assert_type
 from cuda.core.experimental._utils.cuda_utils import (
     _handle_boolean_option,
     check_or_create_options,
