+++ cuda_core/cuda/core/experimental/_stream.py
@@ -13,9 +13,9 @@ from typing import TYPE_CHECKING, Optional, Tuple, Union
 if TYPE_CHECKING:
     import cuda.bindings
     from cuda.core.experimental._device import Device
-from cuda.core.experimental._clear_error_support import assert_type
 from cuda.core.experimental._context import Context
 from cuda.core.experimental._event import Event, EventOptions
+from cuda.core.experimental._utils.clear_error_support import assert_type
 from cuda.core.experimental._utils.cuda_utils import (
     check_or_create_options,
     driver,
